test(server): trim expert-split test surface
