+++ server/CMakeLists.txt
@@ -753,14 +753,6 @@ if(DFLASH27B_TESTS)
         endif()
         target_link_libraries(test_expert_split_runtime PRIVATE dflash_common)
     endif()
-    if(EXISTS "${CMAKE_CURRENT_SOURCE_DIR}/test/test_expert_split_state.cpp")
-        add_executable(test_expert_split_state test/test_expert_split_state.cpp)
-        target_include_directories(test_expert_split_state PRIVATE ${DFLASH27B_SRC_INCLUDE_DIRS})
-        if(DFLASH27B_GPU_BACKEND STREQUAL "cuda")
-            target_include_directories(test_expert_split_state PRIVATE ${CUDAToolkit_INCLUDE_DIRS})
-        endif()
-        target_link_libraries(test_expert_split_state PRIVATE dflash_common)
-    endif()
     if(EXISTS "${CMAKE_CURRENT_SOURCE_DIR}/test/test_expert_split_target_config.cpp")
         add_executable(test_expert_split_target_config test/test_expert_split_target_config.cpp)
         target_include_directories(test_expert_split_target_config PRIVATE ${DFLASH27B_SRC_INCLUDE_DIRS})
@@ -768,14 +760,6 @@ if(DFLASH27B_TESTS)
             target_include_directories(test_expert_split_target_config PRIVATE ${CUDAToolkit_INCLUDE_DIRS})
         endif()
         target_link_libraries(test_expert_split_target_config PRIVATE dflash_common)
-    endif()
-    if(EXISTS "${CMAKE_CURRENT_SOURCE_DIR}/test/test_expert_split_materialization.cpp")
-        add_executable(test_expert_split_materialization test/test_expert_split_materialization.cpp)
-        target_include_directories(test_expert_split_materialization PRIVATE ${DFLASH27B_SRC_INCLUDE_DIRS})
-        if(DFLASH27B_GPU_BACKEND STREQUAL "cuda")
-            target_include_directories(test_expert_split_materialization PRIVATE ${CUDAToolkit_INCLUDE_DIRS})
-        endif()
-        target_link_libraries(test_expert_split_materialization PRIVATE dflash_common)
     endif()
     if(EXISTS "${CMAKE_CURRENT_SOURCE_DIR}/test/test_moe_expert_compute_multi_target.cpp")
         add_executable(test_moe_expert_compute_multi_target test/test_moe_expert_compute_multi_target.cpp)
